Organizando controladores para los modelos
