+++ src/controllers/home_controller.py
@@ -1,0 +1,8 @@
+from src.app import app
+from flask import render_template
+from flask_controller import FlaskController
+
+class HomeController(FlaskController):
+    @app.route('/')
+    def index():
+        return render_template('principal.html', title_page = 'SFI - Home')
